+++ HISTORY.rst
@@ -6,6 +6,7 @@ HEAD
 ----
 
 * BREAKING CHANGE: resource type DateTime is parsed into an object that exposes Kind value
+* FEATURE: access GUIDS stream like a 0-based sequence. GuidHeap subclasses abc.Sequence
 
 0.15.1 (2024)
 -------------